New structure so it supports Jupyter too
