+++ src/github/edit/containerLogic.ts
@@ -1,8 +1,7 @@
 import {
-  addTooltipToHighlight,
-  getHighlightDimensions,
   runCodeValidation,
 } from "../../content_script";
+import { addTooltipToHighlight, getHighlightDimensions } from "../../content_scripts_common";
 import CodigaElement from "../../customelements/CodigaElement";
 import CodigaExtension from "../../customelements/CodigaExtension";
 import CodigaExtensionHighLights from "../../customelements/CodigaExtensionHighlights";
